JC-787 increase timeout for what is considered a fresh data package
we don’t do any closed loop control of the robot.. We run URScripts on the robot and the UR-pc does the closed loop control of the joints. The driver only reads the status of this process. This means that we can safely live with a way higher timeout than the general user of this driver. We already have our own RTDE client that has a timeout of 2000ms. This fix prevents disconnects when we have an slow / unreliable connnection to the robot.