URServer: JC-1093 Specify local ip for tcp-connection
The connection was previously established with an empty string, corresponding to ip 0.0.0.0. It seems like this sometimes can lead to connection issues if the system is run on a pc with more than one network card. Therefore we specify the local ip address to be the same as for the already established rtde-connection.